Factor out accesses to op-regions
The logic to do proper field accesses in all cases is going to get a fair
bit more complex, so I thought it was worth factoring out the actual
field access so it can be accessed from the context. If we need to be able
to override those accesses at any point, that's a better place to have it
anyways.
